+++ blocks/init/src/blocks/assets/scripts/application-blocks-editor.js
@@ -13,19 +13,27 @@
  
  */
 
-import { registerBlocks } from '@eightshift/frontend-libs/scripts/editor';
-import { dynamicImport } from '@eightshift/frontend-libs/scripts/helpers';
+import { registerBlocks, registerVariations } from '@eightshift/frontend-libs/scripts/editor';
 import { Wrapper } from './../../wrapper/wrapper';
-import blocksSettings from './../../manifest.json';
+import WrapperManifest from './../../wrapper/manifest.json';
+import globalSettings from './../../manifest.json';
+import { hooks } from './../../wrapper/wrapper-hooks';
 
 registerBlocks(
-  require.context('./../../custom', true, /manifest\.json$/),
-  require.context('./../../custom', true, /-block.js$/),
-  blocksSettings,
+  globalSettings,
   Wrapper,
-  require.context('./../../custom', true, /transforms.js$/),
-  require.context('./../../custom', true, /icons.js$/),
+  WrapperManifest,
+  require.context('./../../custom', true, /manifest.json$/),
+  require.context('./../../custom', true, /-block.js$/),
+  require.context('./../../custom', true, /-hooks.js$/),
+  require.context('./../../custom', true, /-transforms.js$/),
 );
 
-// Find all blocks hooks require hooks index.js inside it.
-dynamicImport(require.context('./../../wrapper', true, /hooks\/index.js$/));
+registerVariations(
+  globalSettings,
+  require.context('./../../variations', true, /manifest.json$/),
+  require.context('./../../variations', true, /-transforms.js$/),
+);
+
+// Run Wrapper hooks.
+hooks();