+++ blocks/init/src/blocks/custom/button/button-transforms.js
@@ -1,9 +1,10 @@
+import React from 'react'; // eslint-disable-line no-unused-vars
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestButton from './manifest.json';
-import manifestLink from './../link/manifest.json';
+import manifestLink from '../link/manifest.json';
 
-export const Transforms = {
+export const transforms = {
   from: [
     {
       type: 'block',