+++ blocks/init/src/blocks/custom/carousel-image/carousel-image-transforms.js
@@ -1,9 +1,9 @@
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestCarouselImage from './manifest.json';
-import manifestImage from './../image/manifest.json';
+import manifestImage from '../image/manifest.json';
 
-export const Transforms = {
+export const transforms = {
   from: [
     {
       type: 'block',