+++ blocks/init/src/blocks/custom/image/image-transforms.js
@@ -1,9 +1,10 @@
+import React from 'react'; // eslint-disable-line no-unused-vars
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestImage from './manifest.json';
-import manifestCarouselImage from './../carousel-image/manifest.json';
+import manifestCarouselImage from '../carousel-image/manifest.json';
 
-export const Transforms = {
+export const transforms = {
   from: [
     {
       type: 'block',