+++ blocks/init/src/blocks/custom/paragraph/paragraph-transforms.js
@@ -1,9 +1,10 @@
+import React from 'react'; // eslint-disable-line no-unused-vars
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestParagraph from './manifest.json';
-import manifestHeading from './../heading/manifest.json';
+import manifestHeading from '../heading/manifest.json';
 
-export const Transforms = {
+export const transforms = {
   from: [
     {
       type: 'block',